pyproject.toml
[tool.poetry]
name = "fictional-octo-engine"
version = "0.4.0"
description = "A simple Flask Hello World application"
authors = ["Leandro E. Colombo Viña <lecovi@ac.python.org.ar>"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.7"
flask = "^2.2.1"
python-dotenv = "^0.20.0"
flask-redis = "^0.4.0"

[tool.poetry.dev-dependencies]
httpie = "^3.2.1"
pytest-flask = "^1.2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
